// admin/types.d.ts
export interface TradeHistoryType {
  amount: number;
  buyOrSell: string;
  close: string;
  createdAt: string;
  exchangeType: string;
  exchangeTypeIcon: string;
  expireTime: number;
  isProcessed: boolean;
  longOrShortUnit: string;
  open: string;
  price: number;
  profitOrLossAmount: number;
  risk: string;
  riskPercentage: string;
  roi: string;
  status: string;
  symbols: string;
  ticks: string;
  tradeFrom: string;
  tradingMode: string;
  type: string;
  units: number;
  _id: string;
}